refactor(engine): rename worktree-allocator to worktree-plan to end the miner filename collision
gittensory-engine's src/miner/worktree-allocator.ts (pure git-worktree path/branch
planning + retention policy) and gittensory-miner's lib/worktree-allocator.js
(SQLite-backed slot leasing + PID-liveness orphan reclaim) shared a filename while
doing genuinely different jobs. Renames the engine-side PURE planner to
worktree-plan.ts (its export is planWorktree), keeping the miner's allocator name
since it is the one that actually allocates/leases slots (#5668).

Pure rename, no behavior change: exported symbols (planWorktree, WorktreePlan, etc.)
are unchanged, so every @loopover/engine consumer is unaffected; only the engine's
own internal import (worktree-pool.ts) + index re-export path change, plus the two
comment references. Test file renamed to match.

Closes #5668
